Added the StringExtensions.cs file with a public class.
